Move config to Config folder and rename CrdtConfig to HarmonyConfig.
Split builders into separate files under SIL.Harmony.Config and update all references.
