+++ src/SIL.Harmony.Sample/SampleDbContext.cs
@@ -1,11 +1,12 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 using SIL.Harmony.Changes;
+using SIL.Harmony.Config;
 using SIL.Harmony.Db;
 
 namespace SIL.Harmony.Sample;
 
-public class SampleDbContext(DbContextOptions<SampleDbContext> options, IOptions<CrdtConfig> crdtConfig) : DbContext(options), ICrdtDbContext
+public class SampleDbContext(DbContextOptions<SampleDbContext> options, IOptions<HarmonyConfig> crdtConfig) : DbContext(options), ICrdtDbContext
 {
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
