+++ src/SIL.Harmony/Adapters/CustomAdapterProvider.cs
@@ -1,5 +1,6 @@
 using System.Text.Json.Serialization.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using SIL.Harmony.Config;
 using SIL.Harmony.Entities;
 using SIL.Harmony.Helpers;
