+++ src/SIL.Harmony/Config/HarmonyConfig.cs
@@ -1,19 +1,15 @@
 using System.Text.Json;
 using System.Text.Json.Serialization.Metadata;
 using Microsoft.EntityFrameworkCore;
-using SIL.Harmony.Adapters;
 using SIL.Harmony.Changes;
 using SIL.Harmony.Db;
-using SIL.Harmony.Entities;
 using SIL.Harmony.Resource;
 
-namespace SIL.Harmony;
+namespace SIL.Harmony.Config;
 
 public delegate ValueTask BeforeSaveObjectDelegate(object obj, ObjectSnapshot snapshot);
 
-public readonly record struct RegisteredChangeType(Type Type, string Discriminator);
-
-public class CrdtConfig
+public class HarmonyConfig
 {
     /// <summary>
     /// recommended to increase query performance, as getting objects can just query the table for that object.
@@ -34,7 +30,7 @@ public class CrdtConfig
     private readonly Lazy<JsonSerializerOptions> _lazyJsonSerializerOptions;
     private readonly Lazy<ChangeDiscriminatorMaps> _lazyChangeDiscriminatorMaps;
 
-    public CrdtConfig()
+    public HarmonyConfig()
     {
         _lazyChangeDiscriminatorMaps = new Lazy<ChangeDiscriminatorMaps>(BuildChangeDiscriminatorMaps);
         _lazyJsonSerializerOptions = new Lazy<JsonSerializerOptions>(CreateJsonSerializerOptions);
@@ -166,144 +162,3 @@ public class CrdtConfig
         });
     }
 }
-
-internal class JsonOptionsBuilder
-{
-    private bool _frozen;
-    private readonly List<Action<JsonSerializerOptions>> _configurations = [];
-
-    public void Configure(Action<JsonSerializerOptions> configure)
-    {
-        CheckFrozen();
-        _configurations.Add(configure);
-    }
-
-    internal void ApplyTo(JsonSerializerOptions options)
-    {
-        foreach (var configure in _configurations)
-            configure(options);
-        Freeze();
-    }
-
-    private void Freeze() => _frozen = true;
-
-    private void CheckFrozen()
-    {
-        if (_frozen) throw new InvalidOperationException($"{nameof(JsonOptionsBuilder)} is frozen");
-    }
-}
-
-public class ChangeTypeListBuilder
-{
-    private bool _frozen;
-
-    /// <summary>
-    /// we call freeze when the builder is used to create a json serializer options, as it is not possible to add new types after that.
-    /// </summary>
-    public void Freeze()
-    {
-        _frozen = true;
-    }
-
-    private void CheckFrozen()
-    {
-        if (_frozen) throw new InvalidOperationException($"{nameof(ChangeTypeListBuilder)} is frozen");
-    }
-    private readonly List<RegisteredChangeType> _types = [];
-    public IReadOnlyList<RegisteredChangeType> Types => _types.AsReadOnly();
-
-    public ChangeTypeListBuilder Add<TDerived>() where TDerived : IChange, IPolyType
-    {
-        CheckFrozen();
-        if (_types.Any(t => t.Type == typeof(TDerived))) return this;
-        _types.Add(new RegisteredChangeType(typeof(TDerived), TDerived.TypeName));
-        return this;
-    }
-}
-
-public class ObjectTypeListBuilder
-{
-    private bool _frozen;
-
-    /// <summary>
-    /// we call freeze when the builder is used to create a json serializer options, as it is not possible to add new types after that.
-    /// </summary>
-    public void Freeze()
-    {
-        if (_frozen) return;
-        _frozen = true;
-        foreach (var registration in AdapterProviders.SelectMany(a => a.GetRegistrations()))
-        {
-            ModelConfigurations.Add((builder, config) =>
-            {
-                if (!config.EnableProjectedTables) return;
-                var entity = registration.EntityBuilder(builder);
-                entity.HasOne(typeof(ObjectSnapshot))
-                    .WithOne()
-                    .HasForeignKey(registration.ObjectDbType, ObjectSnapshot.ShadowRefName)
-                    .OnDelete(DeleteBehavior.SetNull);
-            });
-        }
-    }
-
-    internal void CheckFrozen()
-    {
-        if (_frozen) throw new InvalidOperationException($"{nameof(ObjectTypeListBuilder)} is frozen");
-    }
-
-    internal Dictionary<Type, List<JsonDerivedType>> JsonTypes { get; } = [];
-    internal List<Action<ModelBuilder, CrdtConfig>> ModelConfigurations { get; } = [];
-    internal List<IObjectAdapterProvider> AdapterProviders { get; } = [];
-
-    public DefaultAdapterProvider DefaultAdapter()
-    {
-        CheckFrozen();
-        if (AdapterProviders.OfType<DefaultAdapterProvider>().SingleOrDefault() is { } adapter) return adapter;
-        adapter = new DefaultAdapterProvider(this);
-        AdapterProviders.Add(adapter);
-        return adapter;
-    }
-
-    /// <summary>
-    /// add a custom adapter for a common interface
-    /// this is required as CRDT objects must express their references and have an Id property
-    /// using a custom adapter allows your model to not take a dependency on Harmony
-    /// </summary>
-    /// <typeparam name="TCommonInterface">
-    /// A common interface that all objects in your application implement
-    /// which System.Text.Json will deserialize your objects to, they must support polymorphic deserialization
-    /// </typeparam>
-    /// <typeparam name="TAdapter">
-    /// This adapter will be serialized and stored in the database,
-    /// it should include the object it is adapting otherwise Harmony will not work
-    /// </typeparam>
-    /// <returns></returns>
-    /// <exception cref="InvalidOperationException">when another adapter has already been added or the config has been frozen</exception>
-    public CustomAdapterProvider<TCommonInterface, TAdapter> CustomAdapter<TCommonInterface, TAdapter>()
-        where TCommonInterface : class where TAdapter : class, ICustomAdapter<TAdapter, TCommonInterface>, IPolyType
-    {
-        CheckFrozen();
-        if (AdapterProviders.OfType<CustomAdapterProvider<TCommonInterface, TAdapter>>().SingleOrDefault() is { } adapter) return adapter;
-        adapter = new CustomAdapterProvider<TCommonInterface, TAdapter>(this);
-        AdapterProviders.Add(adapter);
-        return adapter;
-    }
-
-    internal IObjectBase Adapt(object obj)
-    {
-        if (AdapterProviders is [{ } defaultAdapter])
-        {
-            return defaultAdapter.Adapt(obj);
-        }
-
-        foreach (var objectAdapterProvider in AdapterProviders)
-        {
-            if (objectAdapterProvider.CanAdapt(obj))
-            {
-                return objectAdapterProvider.Adapt(obj);
-            }
-        }
-        throw new ArgumentException($"Unable to adapt object of type {obj.GetType()}");
-    }
-}
-
